Decided on project layout/changed fonts
For project layout: switched from bootstrap cards to large image links with desc on right side, fills space better and is more appealing.
For fonts: learned how to use @font-face in CSS and how to customize fonts in SCSS.
That's one goal done from a previous deployment. Still want to figure out a scroll in animation for project links rather than a load in animation.