Breakpoint spacing fixed/Concact icon
Added an icon to Contact to link to top of page. Fixed all the spacing issues I was having across breakpoints in both the Projects and Intro sections. Settled on About Me section font size and text color. All that's left is replacing all placeholder text, replacing 'logo' with my signature, adding background picture to Intro section, and adding in projects. Website updates may slow down over the next two weeks as I work on a few of them. My 'to-do' list for the site is growing to a close thankfully!